+++ libs/components/a11y/src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/modules/skip-link/skip-link-args';
-export * from './lib/modules/skip-link/skip-link.module';
-export * from './lib/modules/skip-link/skip-link.service';
-export * from './lib/modules/skip-link/skip-link';
+export { SkySkipLinkArgs } from './lib/modules/skip-link/skip-link-args';
+export { SkySkipLinkModule } from './lib/modules/skip-link/skip-link.module';
+export { SkySkipLinkService } from './lib/modules/skip-link/skip-link.service';
+export { SkySkipLink } from './lib/modules/skip-link/skip-link';