+++ libs/components/angular-tree-component/src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/modules/angular-tree/angular-tree.module';
+export { SkyAngularTreeModule } from './lib/modules/angular-tree/angular-tree.module';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.