+++ libs/components/flyout/src/index.ts
@@ -1,13 +1,13 @@
-export * from './lib/modules/flyout/flyout-instance';
-export * from './lib/modules/flyout/flyout.module';
-export * from './lib/modules/flyout/flyout.service';
-export * from './lib/modules/flyout/types/flyout-action';
-export * from './lib/modules/flyout/types/flyout-before-close-handler';
-export * from './lib/modules/flyout/types/flyout-close-args';
-export * from './lib/modules/flyout/types/flyout-config';
-export * from './lib/modules/flyout/types/flyout-message';
-export * from './lib/modules/flyout/types/flyout-message-type';
-export * from './lib/modules/flyout/types/flyout-permalink';
+export { SkyFlyoutInstance } from './lib/modules/flyout/flyout-instance';
+export { SkyFlyoutModule } from './lib/modules/flyout/flyout.module';
+export { SkyFlyoutService } from './lib/modules/flyout/flyout.service';
+export { SkyFlyoutAction } from './lib/modules/flyout/types/flyout-action';
+export { SkyFlyoutBeforeCloseHandler } from './lib/modules/flyout/types/flyout-before-close-handler';
+export { SkyFlyoutCloseArgs } from './lib/modules/flyout/types/flyout-close-args';
+export { SkyFlyoutConfig } from './lib/modules/flyout/types/flyout-config';
+export { SkyFlyoutMessage } from './lib/modules/flyout/types/flyout-message';
+export { SkyFlyoutMessageType } from './lib/modules/flyout/types/flyout-message-type';
+export { SkyFlyoutPermalink } from './lib/modules/flyout/types/flyout-permalink';
 
 // Components and directives must be exported to support Angular’s “partial” Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.