+++ libs/components/forms/src/index.ts
@@ -1,31 +1,31 @@
-export * from './lib/modules/character-counter/character-counter.module';
+export { SkyCharacterCounterModule } from './lib/modules/character-counter/character-counter.module';
 
-export * from './lib/modules/checkbox/checkbox-change';
-export * from './lib/modules/checkbox/checkbox.module';
+export { SkyCheckboxChange } from './lib/modules/checkbox/checkbox-change';
+export { SkyCheckboxModule } from './lib/modules/checkbox/checkbox.module';
 
-export * from './lib/modules/file-attachment/types/file-attachment-change';
-export * from './lib/modules/file-attachment/types/file-attachment-click';
-export * from './lib/modules/file-attachment/types/file-drop-change';
-export * from './lib/modules/file-attachment/file-attachments.module';
-export * from './lib/modules/file-attachment/file-item';
-export * from './lib/modules/file-attachment/file-item-error-type';
-export * from './lib/modules/file-attachment/file-link';
-export * from './lib/modules/file-attachment/file-size.pipe';
+export { SkyFileAttachmentChange } from './lib/modules/file-attachment/types/file-attachment-change';
+export { SkyFileAttachmentClick } from './lib/modules/file-attachment/types/file-attachment-click';
+export { SkyFileDropChange } from './lib/modules/file-attachment/types/file-drop-change';
+export { SkyFileAttachmentsModule } from './lib/modules/file-attachment/file-attachments.module';
+export { SkyFileItem } from './lib/modules/file-attachment/file-item';
+export { SkyFileItemErrorType } from './lib/modules/file-attachment/file-item-error-type';
+export { SkyFileLink } from './lib/modules/file-attachment/file-link';
+export { SkyFileSizePipe } from './lib/modules/file-attachment/file-size.pipe';
 
-export * from './lib/modules/input-box/input-box.module';
-export * from './lib/modules/input-box/input-box-host.service';
-export * from './lib/modules/input-box/input-box-populate-args';
+export { SkyInputBoxModule } from './lib/modules/input-box/input-box.module';
+export { SkyInputBoxHostService } from './lib/modules/input-box/input-box-host.service';
+export { SkyInputBoxPopulateArgs } from './lib/modules/input-box/input-box-populate-args';
 
-export * from './lib/modules/radio/types/radio-change';
-export * from './lib/modules/radio/types/radio-type';
-export * from './lib/modules/radio/radio.module';
+export { SkyRadioChange } from './lib/modules/radio/types/radio-change';
+export { SkyRadioType } from './lib/modules/radio/types/radio-type';
+export { SkyRadioModule } from './lib/modules/radio/radio.module';
 
-export * from './lib/modules/selection-box/types/selection-box-grid-align-items';
-export * from './lib/modules/selection-box/types/selection-box-grid-align-items-type';
-export * from './lib/modules/selection-box/selection-box.module';
+export { SkySelectionBoxGridAlignItems } from './lib/modules/selection-box/types/selection-box-grid-align-items';
+export { SkySelectionBoxGridAlignItemsType } from './lib/modules/selection-box/types/selection-box-grid-align-items-type';
+export { SkySelectionBoxModule } from './lib/modules/selection-box/selection-box.module';
 
-export * from './lib/modules/toggle-switch/types/toggle-switch-change';
-export * from './lib/modules/toggle-switch/toggle-switch.module';
+export { SkyToggleSwitchChange } from './lib/modules/toggle-switch/types/toggle-switch-change';
+export { SkyToggleSwitchModule } from './lib/modules/toggle-switch/toggle-switch.module';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.