+++ libs/components/packages/src/schematics/migrations/update-9/add-forms-popovers-peer-dependency/add-forms-popovers-peer-dependency.schematic.ts
@@ -1,0 +1,14 @@
+import { Rule } from '@angular-devkit/schematics';
+import { NodeDependencyType } from '@schematics/angular/utility/dependencies';
+
+import { ensurePeersInstalled } from '../../../rules/ensure-peers-installed';
+
+export default function (): Rule {
+  return ensurePeersInstalled('@skyux/forms', [
+    {
+      matchVersion: true,
+      name: '@skyux/popovers',
+      type: NodeDependencyType.Default,
+    },
+  ]);
+}