+++ libs/components/pages/src/index.ts
@@ -1,15 +1,18 @@
-export * from './lib/modules/action-hub/action-hub.module';
-export * from './lib/modules/action-hub/types/action-hub-needs-attention';
-export * from './lib/modules/action-hub/types/action-hub-needs-attention-click-handler';
-export * from './lib/modules/action-hub/types/page-link';
-export * from './lib/modules/action-hub/types/page-links-input';
-export * from './lib/modules/action-hub/types/page-modal-link';
-export * from './lib/modules/action-hub/types/page-modal-links-input';
-export * from './lib/modules/action-hub/types/recent-link';
-export * from './lib/modules/action-hub/types/recent-links-input';
-export * from './lib/modules/page/page.module';
-export * from './lib/modules/page/types/page-layout-type';
-export * from './lib/modules/page-header/page-header.module';
+export { SkyActionHubModule } from './lib/modules/action-hub/action-hub.module';
+export { SkyActionHubNeedsAttention } from './lib/modules/action-hub/types/action-hub-needs-attention';
+export {
+  SkyActionHubNeedsAttentionClickHandler,
+  SkyActionHubNeedsAttentionClickHandlerArgs,
+} from './lib/modules/action-hub/types/action-hub-needs-attention-click-handler';
+export { SkyPageLink } from './lib/modules/action-hub/types/page-link';
+export { SkyPageLinksInput } from './lib/modules/action-hub/types/page-links-input';
+export { SkyPageModalLink } from './lib/modules/action-hub/types/page-modal-link';
+export { SkyPageModalLinksInput } from './lib/modules/action-hub/types/page-modal-links-input';
+export { SkyRecentLink } from './lib/modules/action-hub/types/recent-link';
+export { SkyRecentLinksInput } from './lib/modules/action-hub/types/recent-links-input';
+export { SkyPageModule } from './lib/modules/page/page.module';
+export { SkyPageLayoutType } from './lib/modules/page/types/page-layout-type';
+export { SkyPageHeaderModule } from './lib/modules/page-header/page-header.module';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.