+++ libs/components/select-field/src/index.ts
@@ -1,8 +1,8 @@
-export * from './lib/modules/select-field/select-field-picker-context';
-export * from './lib/modules/select-field/select-field.module';
-export * from './lib/modules/select-field/types/select-field';
-export * from './lib/modules/select-field/types/select-field-custom-picker';
-export * from './lib/modules/select-field/types/select-field-select-mode';
+export { SkySelectFieldPickerContext } from './lib/modules/select-field/select-field-picker-context';
+export { SkySelectFieldModule } from './lib/modules/select-field/select-field.module';
+export { SkySelectField } from './lib/modules/select-field/types/select-field';
+export { SkySelectFieldCustomPicker } from './lib/modules/select-field/types/select-field-custom-picker';
+export { SkySelectFieldSelectMode } from './lib/modules/select-field/types/select-field-select-mode';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of public API.